the esm bundle now supports ie11 again
anyone using a bundler will now get ie11 support back, and those that don't support it will get larger bundles.

to help mitigate the bundle size for those that don't need to support ie11, and for those that use single-spa as an in-browser module, an additional bundle called es2015 has been created to fit that need.